[expo-sensors] Pedometer: replace custom listener return type with unimodules Subscription (#12497)
Refs: #12496

During work on the PR above I have spotted that `Pedometer` uses custom listener return type instead of unimodules `Subscription`.

I'm not sure if this refactor is correct, but I have decided to push it anyway to make it easier to see the changes and provide the feedback.

This PR removes the `PedometerListener` custom interface and replaces it by unimodules `Subscription` type (just like the `Battery` module code does).

To add the description to the `Subscription` type I have also had to edit and regenerate the `EventEmitter` file (and add description in there). This change also will be required for the `Battery` module conversion.

CC @brentvatne  @bbarthec

The changes have been tested on `localhost` using regnerated data for `Pedometer` module which is already using the auto-generated API feature.

Correct type of `remove` function will appear automatically after merge of #12496.

<img width="1501" alt="Screenshot 2021-04-12 at 20 05 47" src="https://user-images.githubusercontent.com/719641/114440563-9591ee00-9bca-11eb-8b82-27901e26e67f.png">