[android][web-browser] Add option to not open browser in new task (#12462)
# Why

The addition of this option can improve the user experience for users of apps using this dependency. By not navigating "outside" of the app but instead keeping the users within the app will provide a more seamless experience. I found an issue addressing this [here](https://github.com/expo/expo/issues/11963). This addition will also make the user experience for iOS and Android similar, as the iOS implementation already does this by default.

# How

I created an additional browser option for Android called `createTask` which prevents the addition of the flag 

This resulted in the expected behaviour of the feature and it still maintained the capability of linking back to the application in case of authentication flow.

The default value of this option is set to `true`, which essentially will act in the same way prior to this feature.

# Test Plan

I primarily used the bare-expo app on an emulator running Chrome to be as close as possible to simulate how it would work in practice.

In the following demo I show how it works in the bare-expo app using both `openBrowserAsync()` and `openAuthSessionAsync`.

![demo](https://user-images.githubusercontent.com/10698008/114098686-003ce400-98c2-11eb-9455-1b1d8202509d.gif)

I also added an additional test case to the native tests to make sure the intent flags are set correctly.

<img width="592" alt="Screen Shot 2021-04-08 at 23 31 45" src="https://user-images.githubusercontent.com/10698008/114099129-9d981800-98c2-11eb-83f0-d89bd42e477c.png">