[expo-av][web] Fix Video Fullscreen APIs for Safari (#12258)
**Problem**

The Video.onFullscreenUpdate,videoRef.presentFullscreenPlayer, and videoRef.dismissFullscreenPlayer
APIs are broken in Safari on both iOS and macOS. The videoRef.presentFullscreenPlayer API throws the
error Unhandled Rejection (TypeError): element.requestFullscreen is not a function. (In
'element.requestFullscreen()', 'element.requestFullscreen' is undefined). The other APIs are difficult
to test without that API working, but after a little experimentation I was able to determine that they
also are broken (onFullscreenUpdate is never called even when a video is switched to fullscreen and
dismissFullscreenPlayer throws an error similar to presentFullscreenPlayer).

**Solution**

I extracted the code that interacts with the browser's fullscreen APIs into a separate
FullscreenUtils.web.ts module. I added some feature-detection code to these APIs to switch between the standard
fullscreen API (the previous solution) and the non-standard webkit fullscreen API (my addition to fix
fullscreen usage in Safari).

**Testing**

For each of the browsers listed below I manually:
1. Loaded the Video component page in the bare-expo test app.
2. Confirmed that clicking "Open full screen" switched the video into fullscreen
    and logged `onFullscreenUpdate - {fullscreenUpdate: 1}` to the console.
3. Confirmed that clicking the native close fullscreen control switched the
    video out of fullscreen and logged
    `onFullscreenUpdate - {fullscreenUpdate: 3}` to the console.

Browsers:
* Chrome 89.0.4385.0 on Android 11 (emulated)
* Chrome 89.0.4389.90 on macOS
* Firefox 86.0.1 on macOS
* Safari 14.0.3 on macOS
* Safari on iOS 14.4 (simulator + iPhone SE)

For Safari (iOS and macOS) I manually exectued the webkitExitFullScreen API from
console to confirm that it works as expected (there's no control in the test
app to trigger that API).

For IE11 I just manually tested the DOM APIs/events that I selected. I couldn't find an easy way to
test this PR specifically in IE11 because Expo doesn't work in IE11 by default. I was
tempted to just drop the IE11 fullscreen code I came across because it seemed to be
broken anyway, but Chesterton's fence prevented me.

I did not write unit tests for the new WebFullscreenUtils as that would
basically mean just testing the internal implementation details of the utilities
which is not very helpful.

I could not figure out how to get the `Video` e2e tests running for the web
application - I got strange errors when I tried that. As this is a cross browser
issue I'm not really sure running tests in puppeteer would tell us much though.
Ultimately something like BrowserStack is what's needed to test this change, but
that's beyond the scope of this PR.

fixes #12257