Update view-pager.md
Hi. ViewPager docs are currently linked to an outdated (unsupported) version of @react-native-community/viewpager. It also behaves really buggy on Android, follow for info https://github.com/callstack/react-native-pager-view/issues/164 
On the other hand, as per react-native-pager-view docs, it's not that obvious that it's compatible with Expo starting 5.0.x but it is. Which could be confusing.