[expotools] Run some unit tests using bare-expo project (#12620)
# Why

The `expo-dev-menu` and `expo-dev-launcher` aren't included in the expo go project. So we can't run unit tests using this project. To get around this problem, we can dispatch some tests into the bare-expo project. However, we can't do it for all packages, because not all are included in the bare project. For now, we can run unit tests in both expo go and bare expo.

# Test Plan

- run `expotools native-unit-tests --platform android` ✅