[danger] Warn instead of fail when the changelog isn't modified (#12629)
Why
---
"Fail" is the strongest level of warning that Danger supports and makes sense for things like test failures but comes across as overly strong for a missing changelog entry. There are also several legitimate cases where a changelog entry might not be necessary.

How
---
Changed the `fail` call to a `warn` call.

Test Plan
---------
Add a dummy commit after this one that edits a changelog-tracked package without editing the changelog.