[docs] Filter out errors from browser extensions (#12487)
Browser extensions are a common source of errors and can break pages. There is a tradeoff between learning about errors for which we're not responsible and learning about all errors that are seen in the wild.

This is an RFC PR that proposes filtering out errors known to be from extensions. In practice we have not seen these errors affect a large percentage of site visitors. https://sentry.io/organizations/expoio/issues/?project=1526800&query=is%3Aunresolved+extension&statsPeriod=14d