Fix Issue #1853
On android, Expo Calendar throws a fatal error if the device used has another app for calendar management such as the one mentioned in issue #1853, or in my case "Nine - Email & Calendar" if it is used to sync certain calendars (e.g outlook). 
To error thrown is: `Fatal Exception: java.lang.NumberFormatException: For input string: ""`
To avoid this error I added a check to make sure the string is a parsable integer.