Update TSConfig (#12738)
# Why

I haven't created an issue first, since I thought it would be better to go straight to a pull request, since that illustrates much better what I think you (we?) should change. I hope this is okay.

This is about updating the `tsconfig.json` file and the associated documentation because I think there are some - minor - improvement and clarifications that can be made. This is mostly smaller misunderstandings that I actually usually see i `tsconfig` files.

# How

Update the `tsconfig.json` file and the documentation associated with it.

# Test Plan

The configuration only affects the type checker and not the compiled output. I've tested on my React Native projects, notably <https://github.com/janaagaard75/expo-and-typescript>.