[docs] Fix the tsconfig documentation (#12755)
# Why

Update the documentation for `tsconfig.json` so that it is in sync with the actual `tsconfig.json` used. This is a follow up to https://github.com/expo/expo/pull/12738, because I found a small error.

One question though: Why are we including the `"DOM"` library? The code doesn't run in a browser, so objects like `window` and `document` are not available. It turns out that is needed because `alert()` is available. Could we create a small subset of `"DOM"`, so that only the actually available objects and methods would be listed? I think the documentation line below should explain why `"DOM"` is included, but I couldn't figure out what to write.

# How & Test Plan

The changes in here only affect the documentation, so they haven't been tested.