[web][linear-gradient] fix: useMemo instead of useEffect (#12625)
* useMemo instead of useEffect

Unless I'm missing something, the intent of the code was to only recalculate gradient/angle when some of the props change. For that, `useMemo` is the correct way to do it. With `useMemo`, the first render already has the correct result, instead of waiting for the `useEffect` to run. This also fixes SSR.

* Build package