[docs] Update expo-battery docs to use function component
This updates the expo-battery docs to use a function
component instead of a class component. This is how most modern
React users will likely want to use this.

This change also adds a visualization of a battery to the example
Snack, just to make it more visual and fun. The code for this is
hidden from the docs so that it is not a distraction.