[dev-menu][android] Add websocket support (#12983)
# Why

Closes ENG-1145.

# How

Added web socket control support for dev-menu on iOS (very similar to the https://github.com/expo/expo/pull/12172).

No matter where we swap the command handlers there always will be an instance of `JSPackagerClient` that was created before swapping. The only place where you can add custom handlers is in the `ReactInstanceManagerBuilder.setCustomPackagerCommandHandlers`. Unfortunately, we don't have access to this function. What's worst, we can't even add a new installation step. The builder is hidden from the user. 

So we need to swap command handlers in the current `JSPackagerClient` instance too. However, we can't just use the reflection API to access handlers variable inside that object, cause we don't know if it is available. `JSPackagerClient` is created on background thread (see `DevServerHelper.openPackagerConnection`). The final solution is to spin a background task that monitors if the client is present.
   
# Test Plan

- bare-expo ✅