[kernel] Fix `getInitialURL` returns `https://expo.io:443/<team>/<slug>` instead of null in standalone app (#13071)
# Why

Closes ENG-1141.
Fixes https://github.com/expo/expo/issues/11920.

# How

On both platforms, we pass manifest URL to liking module if the initial URL is null. It's not the desired behavior right now. It looks like we used it in the past but not anymore. So I removed the code that was responsible for changing the initial URL. 

# Test Plan

- check if the linking module work in Expo Go on both platforms ✅
- build local shell apps using `et android-shell-app` and `et ios-shell-app` and confirm that `getInitialURL` returns null using a simple app

> Note: local shell apps are not the same as those produced by turtle because I needed to change a few things to compile them. However, the code which is responsible for starting the app should be the same.