[dev-launcher][android] Reduce crashes when user is spamming deep links (#13020)
# Why

Closes ENG-942.

Reduces app crashes when the user is spamming deep links on Android. Unfortunately, we can't completely remove crashes (because of the RN internal implementation), but now it should happen very rarely.

# How

Added a boolean which holds information whether the app is loading or no. The next app can be loaded only if the previous one was loaded. 

# Test Plan

- open bare-expo and spam `adb shell am start -a android.intent.action.VIEW -d "bareexpo://expo-development-client/?url=exp%3A%2F%2F192.168.0.136%3A19000` in the terminal ✅