[docs] add account member_limit to working together doc (#13413)
Why
A prior PR (#7695) limited the amount of members / open pending invitations on an account to be 100 excluding bots. This PR reflects editing our docs to reflect this behavior. A subsequent PR will update the invitation GUI to reflect this limit.

How
Added a sentence describing new behavior to working-together.md.

Test Plan
N/A