[jest-expo]: Breaking: Switch to React Native v0.64.
react-native@0.64.2, the current latest RN v0.64.x, is meant to work
with just one version of React, react@17.0.1 [1].

jest-expo doesn't yet work with react@17.0.1 because of its
react-test-renderer dependency. Some (well, all) versions in the
react-test-renderer@~16.11.0 range don't work with react@17.0.1 [2].

So, bump the react-test-renderer range to include
react-test-renderer@17.0.1, which works with react@17.0.1 [3] and
will therefore work with react-native@0.64.2.

Then run `yarn`, which updates a root-level lockfile, and check that
tests pass with `yarn test`.

This is breaking because switching to RN v0.64 is forced;
react-test-renderer@~17.0.1 doesn't have backwards compatibility
with the React that's used in RN v0.63.

We keep the tilde because of the report, in a21d7abde, of a semver
violation by react-test-renderer. I didn't look further into that
report or check whether there might be a violation this time; I just
wanted to use the same level of caution that was used there.

[1] https://github.com/facebook/react-native/blob/v0.64.2/package.json#L83
[2] https://github.com/facebook/react/blob/8075c8505/packages/react-test-renderer/package.json#L28
[3] https://github.com/facebook/react/blob/eaaf4cbce/packages/react-test-renderer/package.json#L28