CocoaPods: Use `use_modular_headers!`
I'm not sure if this is a good idea, but since I am using `ReactCommon` and `Folly` in `ExpoModulesCore.podspec` we cannot integrate it as a Swift CocoaPods dependency anymore because of the following error on `pod install`:
```
The Swift pod `ExpoModulesCore` depends upon `ReactCommon` and `Folly`, which do not define modules. To opt into those targets generating module maps (which is necessary to import them from Swift when building as static libraries), you may set `use_modular_headers!` globally in your Podfile, or specify `:modular_headers => true` for particular dependencies.
```