[docs] improve linking in `APISection` component (#13748)
# Why

There were few types marked as non-linkable, because automatically generated links were invalid.

# How

In this small PR I have extracted to the separate method and generalized a bit more the links rendering mechanism, which now allows to hardcode links for certain types.

I'm not sure if linking JS global objects to MDN docs is a good idea, looking for feedback on that. 🙂 

# Test Plan

The changes has been tested by running docs website on `localhost`. The test snapshots has to be updated to reflect the changes to rendering logic.