[dev-launcher][android] Force kotlin version (#13778)
# Why

Forces the same version of kotlin in `dev-launcher package.

# How

The Koin uses a different Kotlin version under the hood - `1.5.X`. The AS will generate a warning because of that. To remove it, we force Koin to use our version of Kotlin. It should be safe, cause the are not that many differences between those two versions. Moreover, we will upgrade Kotlin too eventually.

# Test Plan

- bare-expo ✅