[dev-menu][android] Fix menu binds to all keyboard shortcuts (#13794)
# Why

Closes ENG-1603.
Fixes https://github.com/expo/expo/issues/13731

# How

`onKeyUp` is fired when someone is editing the text input. On Android, inputs use `onKeyDown` and this event won't be pipe to the activity.  However, we can't use `onKeyDown` in our implementation, cause we can't override the RN keyboard shortcut in that scenario. So we are stuck with `onKeyUp`. I've added a detection if the keyboard is visible. If it's we ignore all key events.

# Test Plan

- bare-expo ✅