[dev-menu][android] Fix `Only fullscreen opaque activities can request orientation` (#14066)
# Why

Fixes https://github.com/expo/expo/issues/13965.
Closes ENG-1795.

# How

Due to an issue with API 26, we can't request orientation when the activity is translucent. Unfortunately, this isn't an ideal solution, cause on API 26 users may open the dev-menu in landscape orientation. There may be some visual bugs. So in the future, we may want to add support for landscape orientation. For now, I removed crashes. 

# Test Plan

- new project with dev-menu ✅