[firebase-analytics] Update TS typings to allow null user property values (#14105)
* [firebase-analytics] Allow null user props in ts

This commit modifies the TypeScript typings for expo-firebase-analytics, allowing `null` values to be set for user properties. While this is supported and has a documented effect, (setting the value to null removes the user property), the TypeScript types previously only allowed `string` values for user properties. This commit modifies the types for convenience in working with the analytics module.

* [firebase-analytics] Update CHANGELOG.md

* [firebase-analytics] Update build files

* Update Analytics.ts

* [firebase-analytics] Rebuild Analytics.js.map

Co-authored-by: Kyler Jensen <kylerjensen-b23c19@opayq.com>