[bare-expo][ios] enable hermes (#14119)
# Why

propose and demonstrate a way to configure hermes and works well with config-plugin. our previous discussion might be based on environment variables, but that is not ideal for config-plugin and eas-build. we need add custom environment variable support to eas-build to do that.

# How

this pr introduce a json file near Podfile and config-plugin could reliable update this json file.

# Test Plan

1. set `expo.jsEngine` as hermes or jsc in `podfile.properties.json`
2. `pod install` and `yarn ios`