[modules-core] Fix `Cannot read property 'addProxiedListener' of null` on Android. (#14343)
# Why

Fixes https://github.com/expo/expo/issues/14186.
The solution is based on the one proposed by @Yonom.

# How

In the RN 65 internal implementation of the NativeEventEmiter stars to call `addListener/removeListener` on Android. We provided a custom implementation of those methods which is using the `EXReactNativeEventEmitter`. However, this module is only available on iOS. For now, we can safely make those methods NOOP on Android, but in the future, we should create a similar module to the `EXReactNativeEventEmitter` there too. 

# Test Plan

- simple app with RN 65: