[expo-location] Update error message for LocationUnavailableException
When attempting to acquire a position on Android, the fused location provider occasionally returns `false` when this module calls `isLocationAvailable()`. This results in a `LocationUnavailableException` being thrown. We experienced this issue at @evereepay and we had a really hard time diagnosing it. We feel like the error message for this exception is slightly misleading because the underlying error is a little more broad than what is implied by the existing, "Location provider is unavailable." According to the [documentation](https://developers.google.com/android/reference/com/google/android/gms/location/LocationAvailability#isLocationAvailable%28%29), "Failure to determine location may result from a _number of causes_ including disabled location settings or an inability to retrieve sensor data in the device's environment." This commit / PR introduces a wording change that makes the thrown error message sufficiently ambiguous so as to match the ambiguity in the underlying error.