[location] Update error message for LocationUnavailableException (#14539)
# Why

When attempting to acquire a position on Android, the fused location provider occasionally returns `false` when this module calls `isLocationAvailable()`. This results in a `LocationUnavailableException` being thrown. We experienced this issue at @evereepay and we had a really hard time diagnosing it. Others also experienced this issue, (see #14248), which has since been addressed and partially mitigated. However, we feel like the error message for this exception, should it ever be encountered again, is still slightly misleading and should be updated for clarity.

According to the [documentation](https://developers.google.com/android/reference/com/google/android/gms/location/LocationAvailability#isLocationAvailable%28%29), "Failure to determine location may result from a _number of causes_ including disabled location settings or an inability to retrieve sensor data in the device's environment." This PR introduces a minor wording change that, in our opinion, brings the error message more in line with the actual underlying problem.

# How

1-liner; changed message of `LocationUnavailableException`.

# Test Plan

Throw a `LocationUnavailableException`, observe that the error message is different.

# Checklist

- [ ] Documentation is up to date to reflect these changes (eg: https://docs.expo.io and README.md).
- [x] This diff should work correctly for `expo build` (eg: updated `@expo/xdl`). *<--- Not sure how to do this, might need guidance.*
- [x] This diff should work correctly for `expo prebuild` & EAS Build (eg: updated a module plugin).