Update turtle-cli.md
The text suggesting to visit the Standalone docs are not helpful in finding the certs, profiles, etc. In fact, Googling how to find these values essentially leads to a red-herring. I stumbled upon the commands by accident looking for something else entirely. I believe the commands should live in here in this document, too.