[docs] Improve wording about app.json settings for Google Sign-in (#14053)
* Improve wording about iOS setting for app.json

Original wording does not bring attention to the fact that this is an ios-only settings. The only way to conclude that is to read the json path, which is easily ignored in favor of copy/pasting. The words "Reversed client id" don't mean much to an Android developer and can lead to confusion.

* Add comma

Co-authored-by: Brent Vatne <brentvatne@gmail.com>

* Update Usage with Firebase section

Co-authored-by: Brent Vatne <brentvatne@gmail.com>