[print] Fix page-breaks and margins not supported on iOS (#14802)
The PDF snapshot renderer on iOS does not support page-breaks correctly and renders fonts with incorrect sizes. This is because it leverages the snapshot API for creating image snapshots of pages, but this causes elements to be cut in half and break-inside: avoid; to not be respected.

This PR replaces the snapshot renderer with a ViewPrint renderer which leverages the WKWebView.viewPrintFormatter to render the output to a PDF. This causes elements to be nicely rendered to the next page when they don't fit (and respects break-inside: avoid;) and should be the preferred way of rendering PDF content on iOS. Additionally, it also adds support for a margins option, as @page { margins } is not supported on iOS (only works on Android).