[updates] Workaround double bridge initialization (#15019)
# Why

in expo-updates auto setup flow, we will actually initialize two bridge from startup. the first is from AppDelegate and later replaced with the one from EXUpdatesAppDelegate. if user setup some js code from launch, e.g. faceid authentication. there's chance for user to see faceid authentication twice.

# How

invalidate the first bridge as soon as possible. this is a workaround for sdk43. we will propose a better solution for modules to determine who to create the bridge in sdk44.

# Test Plan

1. keep re-launch release build app to make sure the early invalidation did not have crash.
2. `expo publish` and make sure update works.
3.  `SplashScreen.preventAutoHideAsync()` and delayed `SplashScreen.hideAsync()` after 3000ms to make sure splash screen can keep there longer.


Co-authored-by: Tomasz Sapeta <tomasz.sapeta@swmansion.com>