[firebase-analytics] Remove native `setScreenName` method and deprecate `setCurrentScreen` (#14366)
The native setScreenName/setCurrentScreen methods are deprecated and have meanwhile been removed from the native SDKs. This makes it impossible to upgrade to the latest native Firebase SDKs and causes a warning with the current SDKs.

This PR removes the no longer supported native-methods and implements a (temporary) JS version of setCurrentScreen which simply calls logEvent('screen_view', { screen_name: 'foo' }). This method is marked as deprecated and that it will be removed in SDK 45 to allow users some migration time.

Users need to replace setCurrentScreen with logEvent('screen_view', { screen_name: 'foo' }).