[bare-expo] Make Gradle faster (#15120)
# Why

Makes Gradle a little bit faster in the `bare-expo` project.

# How

- Combined `GenerateDynamicMacrosPlugin` and `ExecuteSetupAndroidProject` into one plugin. 
- Allowed Gradle to run in parallel.
- Allowed Gradle to use file watchers to speed up the compilation.
- Exclude some unneeded operations from the generating dynamic macros command. 
- Changed how `setup-android-project.sh` detect RN version. 

You can check https://proandroiddev.com/how-we-reduced-our-gradle-build-times-by-over-80-51f2b6d6b05b to get more information about some of the optimizations I made. 

# Test Plan

Made several tests involving clean and incremental builds. 
The biggest gain is in the clean builds - there are two times faster. The incremental builds are also faster, but here the difference is more subtle. We can save about from 4 to 8 seconds. It is not a lot, but still an improvement.