[sweet][kotlin] Fix incorrect context in view managers after reload (#15164)
# Why

When I was trying to rewrite `expo-linear-gradient`,  I noticed something weird. The view managers weren't in sync with the most recent module registry. And as it turns out, the view managers are only created once, but other modules are created after each reloads.  So we have to synchronize them manually. 

# How

- Save kotlin view managers in `ModuleRegistryAdapter`.
- Updated module holder in saved view managers when new modules are created.
- Changed how we obtain view managers' names. 

# Test Plan

- bare-expo ✅