[bare-expo] Fix start-metro script on Linux (#15087)
# Why

Running `yarn android` as per documentation on a linux distribution (Manjaro/Arch) fails to successfully run the `start-metro.sh` script with the error `open: command not found`. Upon investigation and research i found that indeed there is no such thing as `open` on linux. The closest thing to it it's `xdg-open`, but executing the `commandFile` with that utility would most probably open it with the default text editor instead of actually executing it.

# How

My humble solution is to check if the OS type is not MacOS and spawn a new default terminal (set by the non-standard `TERMINAL` env variable since i don't currently know any better way to get the default terminal application on linux) window with nohup and redirecting all the output to /dev/null so that it doesn't create the `nohup.out` file...

# Test Plan

On linux, following the CONTRIBUTION.md docs the `start-metro.sh` script should fail with the error `open: command not found`. I did NOT test this on windows/WSL yet but i'm fairly sure the error would be the same.

# Checklist

- [x] Documentation is up to date to reflect these changes (CONTRIBUTION.md)
- [x] Check for os in apps/bare-expo/scripts/start-metro.sh and fix on linux

### Personal notes
My first open source PR :)