[dev-client] Try to fix `java.net.BindException` in unit tests on CI (#15286)
# Why

Tries to fix:
```
java.net.BindException at DevLauncherManifestParserTest.kt:23
```

# How

According to the documentation of the start method:
````
Starts the server on the loopback interface for the given port.
Params:
port - the port to listen to, or 0 for any available port. Automated tests should always use port 0 to avoid flakiness when a specific port is unavailable. 
````
We shouldn't bind to the specific port. 

# Test Plan

- run unit tests locally ✅