[bare-code-scanner] Fix only scans on first mount on Android (#15393)
# Why

Fixes https://github.com/expo/expo/issues/15145.
Fixes https://github.com/expo/expo/issues/9619.
Closes ENG-2490.

# How

`barCodeScannerTaskLock` wasn't set to false when the coroutine scope was canceled, because if your try to run something in the context that was canceled, that instruction will be ignored. That's why we need to check if the scope is active before running `launch`. 

# Test Plan

- expo-go using https://snack.expo.dev/@ajhopen/1x-barcode-scan-issue-sdk43 ✅