In authorisation flows where the oath 'client' is a server and will ultimately do the code for token exchange, it is more secure for the code verifier and code challenge to be generated server side, and for the verifier to never be sent to the client.
Currently within AuthRequest it is not possible to only use a pre-generated code challenge, as the code I am changing here checks to see if codeVerifier is null, and if so generates both a new code verifier and a code challenge.

Changing this check to codeChallenge means it is possible to skip a new pair being generated if my auth sever has already sent me a code challenge to use. It can later use the matching verifier which has never been sent out to the client.