[autolinking] Add `podspecPath` option to config (#15578)
# Why

A better implementation of https://github.com/expo/expo/pull/15564.

# How

We don't want to link an RN module that has an `expo-module.config.json` like `reanimated`. That's why we can't just extend a search region. So I've added a new option to the `expo-module config` that indicates where the podspec is located. 

# Test Plan

- run `node --eval "require('expo-modules-autolinking')(process.argv.slice(1))" -- resolve --platform ios` and check if the podspec path changes.