[autolinking] Fix errors from `expo_patch_react_imports!` (#15699)
fix the errors happened in EAS build yesterday.

there are two errors:

1. ` undefined method expo_patch_react_imports!' for #<Pod::Podfile:0x000000012c8336c8>`

this happens when a managed project have `expo@44.0.0` in yarn.lock and not upgraded yet. it's good to know that `expo prebuild` will not update `expo` version to align with the version from bare template. in this case, i wrap the `expo_patch_react_imports!` with begin-rescue to make sure it doesn't break.

2. `CocoaPods does not put pod subgroups inside nested directories`

this happens when there're some other pod with absolute path. e.g.:
```
  pod 'react-native-google-maps', path: File.dirname(`node --print "require.resolve('react-native-maps/package.json')"`)
```
toward this case, i use the `real_path` method and it's be clear.

[reproducible repo](https://github.com/Kudo/sdk442)

```
expo prebuild
npx pod-install
```

for the second issue, reproducible example in the `upgrade-expo` branch.