[autolinking] Fix errors from `expo_patch_react_imports!` (#15699)
# Why

fix the errors happened in EAS build yesterday.

# How

there are two errors:

1. ` undefined method expo_patch_react_imports!' for #<Pod::Podfile:0x000000012c8336c8>`

this happens when a managed project have `expo@44.0.0` in yarn.lock and not upgraded yet. it's good to know that `expo prebuild` will not update `expo` version to align with the version from bare template. in this case, i wrap the `expo_patch_react_imports!` with begin-rescue to make sure it doesn't break.

2. `CocoaPods does not put pod subgroups inside nested directories`

this happens when there're some other pod with absolute path. e.g.:
```
# @generated begin react-native-maps - expo prebuild (DO NOT MODIFY) sync-e9cc66c360abe50bc66d89fffb3c55b034d7d369
  pod 'react-native-google-maps', path: File.dirname(`node --print "require.resolve('react-native-maps/package.json')"`)
# @generated end react-native-maps
```
toward this case, i use the `real_path` method and it's be clear.

# Test Plan

[reproducible repo](https://github.com/Kudo/sdk442)

```
expo prebuild
# remove the `expo_patch_react_imports!` comment in `ios/Podfile`
npx pod-install
```

for the second issue, reproducible example in the `upgrade-expo` branch.