[ios] Fix orientation lock not working on standalone apps (#15715)
# Why

regression of #14649, the new mechanism relies on rootViewController to pass `shouldAutorotate` and `supportedInterfaceOrientations`. however, in standalone apps, the rootViewController - `EXViewController` doesn't pass the receiver to `EXAppViewController`. 

fix #15666 

# How

create a dedicated `EXStandaloneViewController` to pass `shouldAutorotate` and `supportedInterfaceOrientations` to content view controller (`EXAppViewController`)

# Test Plan

```sh
$ et ios-shell-app --action build --type simulator --configuration Release
$ et ios-shell-app --url https://exp.host/@kudochien/expo-mcve --action configure --type simulator -s 44.0.0 --archivePath shellAppBase-simulator/Build/Products/Release-iphonesimulator/ExpoKitApp.app --output app.tar.gz
```
verify the screen orientation is locked to portrait or not.