[eyw] Add .cjs and .mjs to allowed file extensions
Why
---
When we updated `@apollo/client` to a newer minor/patch version due to a yarn.lock refresh, it caused the publishing step for Home to fail because that package specifies a .cjs file under package.json->main. This commit adds .cjs and .mjs to the list of extensions treated as source files.

How
---
Add cjs and mjs to sourceExts for our Metro config.

Test Plan
---------
Open Home in Expo Go and see that it bundles successfully and loads and runs successfully.