[sweet][kotlin] Fix callbacks not working if only one callback was registered (#16065)
# Why

Fix callbacks not working if only one callback was registered.

# How

For some reason, RN requires that the `getExportedCustomDirectEventTypeConstants` will be a hashmap. Kotlin collection factory functions use a different map type if only one element is present. I tested that feature with multiple callbacks registered. So I missed that detail.

# Test Plan

- bare-expo ✅