fix: update `fbemitter` (#16245)
# Why

As discussed in https://github.com/expo/expo/pull/16023#issuecomment-1021732956, `fbemitter@2` pulls in an old version of `fbjs` which in turns pulls in a vulnerable version of `node-fetch` (https://github.com/advisories/GHSA-r683-j2x4-v87g). Upgrading `fbemitter` resolves this.

Note that the repo of `fbemitter` has been archived, so exploring a way to replace it entirely is probably a good idea.

# How

I ran `yarn upgrade-interactive --latest` and picked `fbemitter` from the list (not `native-component-list`, but the three other instances).

# Test Plan

As mentioned in https://github.com/expo/expo/pull/16023#issuecomment-1033668433, there doesn't seem to be any actual diff between the versions except

1. The update of `fbjs` (which is why `expo` should update as well)
2. `var` -> `const`
3. Different formatting rules (such as trailing comma)

So from what I can tell, this is completely safe to land. I haven't (and probably won't unfortunately) had the time to properly build and start this up to test it. But as mentioned above, from my reading of the diff between `2.1.1` and `3.0.0` there's no change to the logic of the code.