[dev-launcher] Add crash report screen (#16341)
# Why

Closes ENG-2740

# How

Saved crash information on iOS (it was done before on Android) and surface them on the next launch. 
In case of a crash, on the main screen, you will see the information that something went wrong and after taping it, you will see a crash report screen.

> Note:
The design isn't final but should be good enough for now.