[expo-localization] fix: Exception in HostObject::get for prop 'NativeUnimoduleProxy': java.lang.NullPointerException (#16316)
# Why

Fixes #15899

I just migrated to the expo from `react-native-unimodules`. this error was happening only for android when I open app on android.
**if I uninstall the app and reinstall the app then this error is gone.**

# How

by tested this code for `expo-localization` https://github.com/expo/expo/issues/15899#issuecomment-1039952877 it is working for me & @photorouletteadmin

# Test Plan

you have to migrate the expo from `react-native-unimodules` to `expo-modules`. this error was happening only for android when you will install the app over existing app and open app.