[docs] Mask permissions when creating SSH priv key
In the Git submodule docs there's a race condition between creating the
SSH private key and setting its permissions.

It's generally better to set the umask so that the permissions are
correct up front rather than having the permissions incorrect initially
until they are changed.

If the example eas-build-pre-install hook is only used on the build
servers it probably won't make a difference, but if someone uses it on a
multi-user system, their private key would be exposed very briefly if
it's done with a create followed by a chmod.