[modules-core][android] Improve test library (#16416)
# Why

Part of ENG-4140.

This PR introduces a new way of mocking an expo module. Currently, the code lives in the `unimodules-core-test` package. I plan to rename that later. 

# How

Existing mocking mechanism forces as to call methods via their names and mock the promise. It works fine, however, we can do it better. Our modules API promotes writing promises-less functions (promises are hidden from users). So why do we have to deal with promises even if the call on the native side is 100% sync? Therefore, I've created an abstraction that uses the invocation handler to remove promises.

## How does it work?

Firstly, you have to create a module test interface that will be used to call functions exported via DSL.  You can think of that as a native interface for js exported functions. 
For instance module like this:
```kt
class MyModule : Module() {
  override fun definition() = ModuleDefinition {
    function("f1") {
      return@function "string"
    }

    function("f2") { param: String ->
      return@function true
    }
  }
}
```
can be typed using an interface like this:
```kt
interface MyModuleTestInterface {
	fun f1(): String
	fun f2(param: String): Bool
}
```
The test interface could be generated in the future, but for now, it is beyond the scope.

After that, you are ready to create a mocked module:
```kt
ModuleMock.createMock(MyModuleTestInterface::class, MyModule()) {
	// here you can access to the mocked object
}
```
In the `createMock` block you have access to the `MyModuleTestInterface` object and if you call something on that it will be delegated to the exported method.

# Test Plan

- run tests for `expo-clipboard`