[sweet api][android] Add validators & required fields to records (#16529)
# Why

Part of https://linear.app/expo/issue/ENG-4123/more-customized-arguments-validation

# How

I've decided to finally add a required field and validators to the records structure. The required field was pretty much straightforward. However, validators were very tricky to implement.  First of all, we have something called `ValidationBinder`.  It's an annotation that you have to add to the final validator annotation to transfer data from annotation to the validator class itself. Let me explain that in more detail using an example:

To create a custom validator: `@IntRange` you will have to define an annotation (that can receive custom arguments) and connect to the binder using `@BindUsing` annotation. The `@BindUsing` requires one argument - class of the binder. The binder implementation will have to convert an annotation object `@IntRange` to an instance of `FieldValidator`. This object is used later to check if the js value is correct. 

# Test Plan

- unit tests ✅