[autolinking][android] Fix `cannot cast object 'ExpoAutolinkingManager@' with class 'ExpoAutolinkingManager' to class 'ExpoAutolinkingManager'` (#16545)
# Why

Closes ENG-4204.

# How

If a project is using `buildSrc`, the Gradle will use different class loaders for `settings.gradle` and for `build.gradle` scripts in modules. Because of that, we can't cast manager created in `settings.gradle` to `ExpoAutolinkingManager` in the `build.gradle` of expo package. I didn't find any workaround, but we don't have to make that cast. 

# Test Plan

- build project with `buildSrc` ✅