feat(cli): add `--check` and `--fix` flags to install command
Add a flag `--check` which runs the doctor check against the installed packages, then prompt to install the corrected packages. If the command is run in CI then no prompt will be given and the process with exit with non-zero exit code (1).

Additionally adds the `--fix` flag which skips the prompt and exiting behavior. This can be used in CI to naively ensure packages are always correct: `npx expo install --fix`.

In this logical scenario `--fix` acts as a more privileged version of `--check`. Meaning the flags should never be used together. I also considered the inverse structure `--fix --dry-run` which would act the same as `--check` but would allow for coherency in the arguments since you would have less of a chance to provide bad arguments.

I opted for the `--check` flag for this initial commit because it allows for the interactive prompt to fix when run outside of CI, and acts as expected in CI. Whereas `--dry-run` exiting with non-zero seems unexpected as running without `--dry-run` would not exit with non-zero (unless for some unrelated error).

This feature also combines all three sources for known versions before doing validation, this means the check run during `start` will utilize the same sources as the install check.

In addition, I've also added `react-native`, `react`, `react-dom`, and `react-native-web` to the `bundledNativeModules.json` so the React versions are all checked as well.

The result is a unified validation system across the CLI.

This feature supersedes the `expo doctor --fix-dependencies` command, while adding the additional ability to check/fix only a certain set of packages, i.e. `npx expo install react-native --check`